backend/requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.15
django-cors-headers>=4.4
celery>=5.3
redis>=5.0
python-dotenv>=1.0
